Working LLM Rag model. Next steps would be to add tensors from sentence chunks to an active database and retrieve them and give security features
